+++ truffle/CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * GR-41408 Added `${graalvm-version}` and `${graalvm-website-version}` substitutions for the `website` property of language and instrument registrations.
 
 * GR-41034 Added `TruffleInstrument.Env.getTruffleFile(TruffleContext, ...)` methods to allow reading a truffle file from a specific context without being entered. Deprecated `TruffleInstrument.Env.getTruffleFile(...)` methods that do not take the `TruffleContext`.
+* GR-42271 Native image build verifies that the context pre-initialization does not introduce absolute TruffleFiles into the image heap. The check can be disabled using the `-H:-TruffleCheckPreinitializedFiles` native image option.
 
 * GR-40274 TruffleStrings: added AsNativeNode and GetStringCompactionLevelNode.